refactor: move template-driven components
