+++ projects/ng-tailwind/src/components/forms/template-driven/ngt-date/ngt-date.module.ts
@@ -3,9 +3,9 @@ import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 
 import { EvDatePickerModule } from 'ev-date-picker';
-import { NgtHelperModule } from '../ngt-helper/ngt-helper.module';
-import { NgtShiningModule } from '../ngt-shining/ngt-shining.module';
-import { NgtValidationModule } from '../ngt-validation/ngt-validation.module';
+import { NgtHelperModule } from '../../../ngt-helper/ngt-helper.module';
+import { NgtShiningModule } from '../../../ngt-shining/ngt-shining.module';
+import { NgtValidationModule } from '../../../ngt-validation/ngt-validation.module';
 import { NgtDateComponent } from './ngt-date.component';
 
 @NgModule({
