+++ projects/ng-tailwind/src/components/forms/template-driven/ngt-input/ngt-input.component.html
