+++ projects/ng-tailwind/src/components/forms/template-driven/ngt-input/ngt-input.module.ts
@@ -3,10 +3,10 @@ import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { NgtFormModule } from '../ngt-form/ngt-form.module';
-import { NgtHelperModule } from '../ngt-helper/ngt-helper.module';
-import { NgtShiningModule } from '../ngt-shining/ngt-shining.module';
-import { NgtSvgModule } from '../ngt-svg/ngt-svg.module';
-import { NgtValidationModule } from '../ngt-validation/ngt-validation.module';
+import { NgtHelperModule } from '../../../ngt-helper/ngt-helper.module';
+import { NgtShiningModule } from '../../../ngt-shining/ngt-shining.module';
+import { NgtSvgModule } from '../../../ngt-svg/ngt-svg.module';
+import { NgtValidationModule } from '../../../ngt-validation/ngt-validation.module';
 import { NgtInputComponent } from './ngt-input.component';
 
 @NgModule({
