+++ projects/ng-tailwind/src/components/forms/template-driven/ngt-multi-select/ngt-multi-select.component.css
