+++ projects/ng-tailwind/src/components/forms/template-driven/ngt-multi-select/ngt-multi-select.module.ts
@@ -2,12 +2,12 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 
-import { NgtStylizableModule } from '../../directives/ngt-stylizable/ngt-stylizable.module';
+import { NgtStylizableModule } from '../../../../directives/ngt-stylizable/ngt-stylizable.module';
 import { NgtCheckboxModule } from '../ngt-checkbox/ngt-checkbox.module';
-import { NgtHelperModule } from '../ngt-helper/ngt-helper.module';
+import { NgtHelperModule } from '../../../ngt-helper/ngt-helper.module';
 import { NgtInputModule } from '../ngt-input/ngt-input.module';
-import { NgtShiningModule } from '../ngt-shining/ngt-shining.module';
-import { NgtValidationModule } from '../ngt-validation/ngt-validation.module';
+import { NgtShiningModule } from '../../../ngt-shining/ngt-shining.module';
+import { NgtValidationModule } from '../../../ngt-validation/ngt-validation.module';
 import { NgtMultiSelectComponent } from './ngt-multi-select.component';
 
 @NgModule({
