+++ projects/ng-tailwind/src/components/forms/template-driven/ngt-radio-button/ngt-radio-button.module.ts
@@ -1,8 +1,8 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
-import { NgtHelperModule } from '../ngt-helper/ngt-helper.module';
-import { NgtShiningModule } from '../ngt-shining/ngt-shining.module';
+import { NgtHelperModule } from '../../../ngt-helper/ngt-helper.module';
+import { NgtShiningModule } from '../../../ngt-shining/ngt-shining.module';
 import { NgtRadioButtonContainerComponent } from './ngt-radio-button-container/ngt-radio-button-container.component';
 import { NgtRadioButtonComponent } from './ngt-radio-button.component';
