+++ projects/ng-tailwind/src/components/forms/template-driven/ngt-select/ngt-select.component.css
