+++ projects/ng-tailwind/src/components/forms/template-driven/ngt-select/ngt-select.module.ts
@@ -4,10 +4,10 @@ import { FormsModule } from '@angular/forms';
 import { NgSelectModule } from '@ng-select/ng-select';
 
 import { NgtFormModule } from '../ngt-form/ngt-form.module';
-import { NgtHelperModule } from '../ngt-helper/ngt-helper.module';
-import { NgtShiningModule } from '../ngt-shining/ngt-shining.module';
-import { NgtSvgModule } from '../ngt-svg/ngt-svg.module';
-import { NgtValidationModule } from '../ngt-validation/ngt-validation.module';
+import { NgtHelperModule } from '../../../ngt-helper/ngt-helper.module';
+import { NgtShiningModule } from '../../../ngt-shining/ngt-shining.module';
+import { NgtSvgModule } from '../../../ngt-svg/ngt-svg.module';
+import { NgtValidationModule } from '../../../ngt-validation/ngt-validation.module';
 import { NgtSelectComponent } from './ngt-select.component';
 import { NgtSelectHeaderTmp, NgtSelectOptionSelectedTmp, NgtSelectOptionTmp } from './ngt-select.directive';
