+++ projects/ng-tailwind/src/components/forms/template-driven/ngt-slider/ngt-slider.component.css
