+++ projects/ng-tailwind/src/components/forms/template-driven/ngt-slider/ngt-slider.module.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
-import { NgtShiningModule } from '../ngt-shining/ngt-shining.module';
+import { NgtShiningModule } from '../../../ngt-shining/ngt-shining.module';
 import { NgtSliderComponent } from './ngt-slider.component';
 
 @NgModule({
