+++ projects/ng-tailwind/src/components/forms/template-driven/ngt-textarea/ngt-textarea.component.html
