+++ projects/ng-tailwind/src/components/forms/template-driven/ngt-textarea/ngt-textarea.module.ts
@@ -3,9 +3,9 @@ import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { NgtFormModule } from '../ngt-form/ngt-form.module';
-import { NgtHelperModule } from '../ngt-helper/ngt-helper.module';
-import { NgtShiningModule } from '../ngt-shining/ngt-shining.module';
-import { NgtValidationModule } from '../ngt-validation/ngt-validation.module';
+import { NgtHelperModule } from '../../../ngt-helper/ngt-helper.module';
+import { NgtShiningModule } from '../../../ngt-shining/ngt-shining.module';
+import { NgtValidationModule } from '../../../ngt-validation/ngt-validation.module';
 import { NgtTextareaComponent } from './ngt-textarea.component';
 
 @NgModule({
